feat: fine tune pyenv loading
Now paths added to $PATH are inside the config.fish file. Allowing for easy modification and changes taking places instantly, as opposed to universal variables which fully take effect after re-logging.

It was also explicitely modified $PATH to prepend the current working working directory in case it is not in it.